+++ scraperx/__init__.py
@@ -105,7 +105,7 @@ from .social_db import SocialDB
 from .thread import Thread, get_thread
 from .token_extractor import TokenMention, extract_token_mentions
 
-__version__ = "1.10.0"
+__version__ = "1.11.0"
 
 __all__ = [
     "AvatarMatcher",